// src/app/shared/models/contact.ts
export class Contact{
    id: string;
    dateEntered: number;
    dateModified: number;
    modifiedUserId: string;
    createdBy: string;
    description: string;
    deleted: number;
    assignedUserId: string;
    assignedUserName: string
    salutation: string;
    firstName: string;
    lastName: string;
    title: string;
    department: string;
    doNotCall: boolean;
    phoneHome: string;
    phoneMobile: string;
    phoneWork: string;
    phoneOther: string;
    phoneFax: string;
    primaryAddressStreet: string;
    primaryAddressCity: string;
    primaryAddressState: string;
    primaryAddressPostalcode: string;
    primaryAddressCountry: string;
    altAddressStreet: string;
    altAddressCity: string;
    altAddressState: string;
    altAddressPostalcode: string;
    altAddressCountry: string;
    assistant: string;
    assistantPhone: string;
    reportsToId: string;
    birthdate: number;
    accountId: string;
    accountName: string;
}